Migrate fjs/djs/serializer to .f.mjs
Implementation and proof migrate from authored TypeScript to
JSDoc-typed JavaScript under the stage-1 migration. MapEntries,
RefCounter, Refs, GetConstsState, and RefLookup stay as _-prefixed
JSDoc typedefs in the module — nothing outside the module imports
them by name. Updates fjs/djs/module.f.ts, fjs/djs/tokenizer/module.f.ts
(still TypeScript, but its runtime dependency on serializer is now
resolved), and five proof files' import specifiers, and fixes stale
module.f.ts/proof.f.ts doc references across six todo docs — most
already had stale line-number citations predating this migration
(verified against the pre-migration file), so only the .f.ts -> .f.mjs
extension was corrected there; two references whose line numbers still
matched were fixed exactly.
