+++ fjs/djs/module.f.ts
@@ -4,7 +4,7 @@
  * @module
  */
 import { transpile } from './transpiler/module.f.ts'
-import { stringify, stringifyAsTree } from './serializer/module.f.ts'
+import { stringify, stringifyAsTree } from './serializer/module.f.mjs'
 import { sort } from '../types/object/module.f.mjs'
 import { pure, step } from '../effects/module.f.mjs'
 import type { Effect } from '../effects/types.ts'
