+++ fjs/djs/todo/157.md
@@ -98,8 +98,8 @@ id/ref hooks, then wraps the result in its module-level
 The same recursive `typeof`-dispatch walker is written three times:
 
 - `fjs/media/json/module.f.ts:49` (`serialize`)
-- `fjs/djs/serializer/module.f.ts:79` (`serializeWithoutConst`)
-- `fjs/djs/serializer/module.f.ts:117` (`serializeWithConst`)
+- `fjs/djs/serializer/module.f.mjs:79` (`serializeWithoutConst`)
+- `fjs/djs/serializer/module.f.mjs:117` (`serializeWithConst`)
 
 Each defines the identical closure cluster: `propertySerialize`
 (`flat([stringSerialize(k), colon, f(v)])`), `mapPropertySerialize`,
