+++ fjs/djs/todo/197.md
@@ -22,10 +22,10 @@ recursion or terminal handling):
 | # | Location | Purpose |
 |---|---|---|
 | 1 | `fjs/media/json/module.f.ts:67` — `serialize.f` | JSON `typeof` switch (a subset: no `bigint` / `undefined`). |
-| 2 | `fjs/djs/serializer/module.f.ts:99` — `serializeWithoutConst.f` | DJS serialize (+`bigint` +`undefined`). |
-| 3 | `fjs/djs/serializer/module.f.ts:135` — `serializeWithConst.f` | (2) + ref-counter short-circuit. |
-| 4 | `fjs/djs/serializer/module.f.ts:36` — `getConstantsOp` | Collect constants for `const c{n} = …` block. |
-| 5 | `fjs/djs/serializer/module.f.ts:163` — `countRefsOp` | Count references for the ref table. |
+| 2 | `fjs/djs/serializer/module.f.mjs:99` — `serializeWithoutConst.f` | DJS serialize (+`bigint` +`undefined`). |
+| 3 | `fjs/djs/serializer/module.f.mjs:135` — `serializeWithConst.f` | (2) + ref-counter short-circuit. |
+| 4 | `fjs/djs/serializer/module.f.mjs:36` — `getConstantsOp` | Collect constants for `const c{n} = …` block. |
+| 5 | `fjs/djs/serializer/module.f.mjs:163` — `countRefsOp` | Count references for the ref table. |
 | 6 | `fjs/djs/ast/module.f.mjs:41` — `toDjs` | Evaluate `AstConst` → `Unknown` (over `AstConst`, a parallel shape with `'aref'`/`'cref'`/`'array'` tuples). |
 
 [i157 §2](todo.md) covers (1)–(3) by factoring the
