+++ fjs/djs/transpiler/proof.f.ts
@@ -1,6 +1,6 @@
 import { sort } from '../../types/object/module.f.mjs'
 import { transpile } from './module.f.ts'
-import { stringifyAsTree } from '../serializer/module.f.ts'
+import { stringifyAsTree } from '../serializer/module.f.mjs'
 import { virtual, emptyState } from '../../effects/node/virtual/module.f.mjs'
 import type { Dir } from '../../effects/node/virtual/types.ts'
 import { utf8 } from '../../text/module.f.mjs'
