+++ fjs/media/json/todo/stringify-sorted-canonical.md
@@ -26,7 +26,7 @@ Proof files (each binds its own alias: `jsonStr`, `str`, `stringify`,
   `fjs/text/utf16/proof.f.ts:15`
 - `fjs/media/json/parser/proof.f.ts:13`, `fjs/protocol/mcp/stdio/proof.f.mjs:20`
 - `fjs/bnf/data/proof.f.ts` (10 inline calls), `fjs/djs/parser/proof.f.ts:306`,
-  `fjs/djs/serializer/proof.f.ts:47`
+  `fjs/djs/serializer/proof.f.mjs:47`
 
 Each site is one line, so no single site is a problem — the issue is that
 the canonical-serialization idiom has ~20 different local names and no
